Dashboard/requirements.txt
absl-py==0.14.1
anyio==3.3.0
appdirs==1.4.4
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==3.3.1
cachetools==4.2.4
category-encoders==2.3.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.8
clang==5.0
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
cycler==0.11.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
dirty-cat==0.2.0
docopt==0.6.2
entrypoints==0.3
Flask==2.0.2
flatbuffers==1.12
fonttools==4.28.2
gast==0.4.0
geographiclib==1.52
geomet==0.3.0
geopy==2.2.0
google-api-core==2.2.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.30.1
google-cloud-core==2.2.1
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.42.0
grpcio-status==1.42.0
h5py==3.1.0
idna==3.3
imageio==2.9.0
importlib-metadata==4.8.1
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.3
joblib==1.0.1
Js2Py==0.71
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-server==1.10.2
jupyterlab==3.1.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.7.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.2
liac-arff==2.5.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.2
mglearn==0.1.9
minio==7.1.0
mistune==0.8.4
munch==2.5.0
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
notebook-as-pdf==0.5.0
numpy==1.21.4
oauthlib==3.1.1
openml==0.12.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.3
pandocfilters==1.4.3
parso==0.8.2
patsy==0.5.2
pickleshare==0.7.5
Pillow==8.4.0
pipwin==0.5.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
proto-plus==1.19.8
protobuf==3.19.1
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyee==8.2.2
Pygments==2.9.0
pyjsparser==2.7.1
pyparsing==3.0.6
PyPDF2==1.26.0
pyppeteer==0.2.6
PyPrind==2.11.3
pyproj==3.2.1
pyrsistent==0.18.0
pySmartDL==1.3.4
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.10.0
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.8
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.7.1
setuptools-scm==6.3.2
Shapely==1.8.0
simplejson==3.17.5
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
statsmodels==0.13.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.2.0
tomli==1.2.2
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
txt2tags==3.7
typing-extensions==3.7.4.3
tzdata==2021.5
tzlocal==4.0.2
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
websockets==9.1
Werkzeug==2.0.2
widgetsnbextension==3.5.1
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.6.0